refactor: move ticket DTOs to ticket package
